+++ packages/owc-dev-server/demo/module-b.js
@@ -1,0 +1,7 @@
+/* eslint-disable */
+import { html } from 'lit-html';
+
+export const foo = () => 'module b foo';
+
+console.log('module b');
+console.log('lit-html', html);