+++ packages/polyfills-loader/src/webcomponents-sd-ce.js
@@ -1,0 +1,3 @@
+import '@webcomponents/shadydom/src/shadydom.js';
+import '@webcomponents/custom-elements/src/custom-elements.js';
+import '@webcomponents/shadycss/entrypoints/scoping-shim.js';