+++ packages/testing-helpers/index-no-side-effects.js
@@ -1,0 +1,14 @@
+export { html, unsafeStatic } from './lit-html.js';
+export {
+  triggerBlurFor,
+  triggerFocusFor,
+  oneEvent,
+  isIE,
+  defineCE,
+  aTimeout,
+  nextFrame,
+} from './helpers.js';
+export { litFixture, litFixtureSync } from './litFixture.js';
+export { stringFixture, stringFixtureSync } from './stringFixture.js';
+export { fixture, fixtureSync } from './fixture-no-side-effect.js';
+export { cachedWrappers, fixtureWrapper, fixtureCleanup } from './fixtureWrapper.js';