+++ packages/testing-helpers/index.js
@@ -11,3 +11,4 @@ export {
 export { litFixture, litFixtureSync } from './litFixture.js';
 export { stringFixture, stringFixtureSync } from './stringFixture.js';
 export { fixture, fixtureSync } from './fixture.js';
+export { cachedWrappers, fixtureWrapper, fixtureCleanup } from './fixtureWrapper.js';