+++ packages/webpack-import-meta-loader/test/caseB/index.js
@@ -1,0 +1,3 @@
+import './caseBsub/caseBsub';
+
+window.foo = new URL('./', import.meta.url);