+++ packages/testing-karma-bs/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.7](https://github.com/open-wc/open-wc/tree/master/packages/testing-karma-bs/compare/@open-wc/testing-karma-bs@0.2.6...@open-wc/testing-karma-bs@0.2.7) (2019-01-26)
+
+
+### Bug Fixes
+
+* align all open-wc readme headers ([b589429](https://github.com/open-wc/open-wc/tree/master/packages/testing-karma-bs/commit/b589429))
+
+
+
+
+
+## [0.2.6](https://github.com/open-wc/open-wc/tree/master/packages/testing-karma-bs/compare/@open-wc/testing-karma-bs@0.2.5...@open-wc/testing-karma-bs@0.2.6) (2019-01-20)
+
+
+### Bug Fixes
+
+* refactor generators ([1dab1f4](https://github.com/open-wc/open-wc/tree/master/packages/testing-karma-bs/commit/1dab1f4))
+
+
+
+
+
+## [0.2.5](https://github.com/open-wc/open-wc/tree/master/packages/testing-karma-bs/compare/@open-wc/testing-karma-bs@0.2.4...@open-wc/testing-karma-bs@0.2.5) (2019-01-19)
+
+**Note:** Version bump only for package @open-wc/testing-karma-bs
+
+
+
+
+
+## [0.2.4](https://github.com/open-wc/open-wc/tree/master/packages/testing-karma-bs/compare/@open-wc/testing-karma-bs@0.2.3...@open-wc/testing-karma-bs@0.2.4) (2019-01-16)
+
+
+### Bug Fixes
+
+* improve documentation ([4f5472f](https://github.com/open-wc/open-wc/tree/master/packages/testing-karma-bs/commit/4f5472f))
+
+
+
+
+
+## [0.2.3](https://github.com/open-wc/open-wc/tree/master/packages/testing-karma-bs/compare/@open-wc/testing-karma-bs@0.2.2...@open-wc/testing-karma-bs@0.2.3) (2019-01-09)
+
+**Note:** Version bump only for package @open-wc/testing-karma-bs
+
+
+
+
+
 ## [0.2.2](https://github.com/open-wc/open-wc/tree/master/packages/testing-karma-bs/compare/@open-wc/testing-karma-bs@0.2.1...@open-wc/testing-karma-bs@0.2.2) (2018-12-20)
 
 **Note:** Version bump only for package @open-wc/testing-karma-bs